Introduce a way for running system tests in parallel.
Run reference test by calling:
$ python log_tests.py TestHWBasedSuccessfulTestruns.test_hw_based_testrun_with_test_definition_tests &

When using SystemTest for triggering the testrun it's possible to run any
number of system tests in parallel. Previously it was possible to run
only one test at a time (running all tests in sequence takes a lot of
time). Also now it doesn't matter if the server is running
some other tests at the same time, SystemTest makes it possible to run
automated system tests against production server for instance.

In the future (if SystemTest is used for everything) all tests can be
run in parallel using:
$ nosetests -s --processes=8 log_tests.py
Number of processes used should be less than the number of devices used.

URL to testrun log is printed to console a few seconds after the testrun is
started, which makes debugging much easier.